backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-cors==4.0.0
requests==2.31.0
openai>=1.12.0
deep-translator==1.11.4
supabase==1.2.0 